refactor: Simplified Bayesian optimization code
