+++ syne_tune/optimizer/schedulers/searchers/__init__.py
@@ -30,6 +30,10 @@ from syne_tune.optimizer.schedulers.searchers.random_grid_searcher import (  # n
 from syne_tune.optimizer.schedulers.searchers.searcher_factory import (  # noqa: F401
     searcher_factory,
 )
+from syne_tune.optimizer.schedulers.searchers.model_based_searcher import (  # noqa: F401
+    ModelBasedSearcher,
+    BayesianOptimizationSearcher,
+)
 
 __all__ = [
     "BaseSearcher",
@@ -40,12 +44,11 @@ __all__ = [
     "RandomSearcher",
     "GridSearcher",
     "searcher_factory",
+    "ModelBasedSearcher",
+    "BayesianOptimizationSearcher",
 ]
 
 try:
-    from syne_tune.optimizer.schedulers.searchers.model_based_searcher import (  # noqa: F401
-        ModelBasedSearcher,
-    )
     from syne_tune.optimizer.schedulers.searchers.gp_fifo_searcher import (  # noqa: F401
         GPFIFOSearcher,
     )
@@ -55,7 +58,6 @@ try:
 
     __all__.extend(
         [
-            "ModelBasedSearcher",
             "GPFIFOSearcher",
             "GPMultiFidelitySearcher",
         ]
