+++ syne_tune/optimizer/schedulers/searchers/model_based_searcher.py
@@ -13,6 +13,8 @@
 import time
 from typing import Optional, Type, Dict, Any, List
 import logging
+import numpy as np
+import copy
 
 from syne_tune.optimizer.schedulers.searchers import (
     StochasticSearcher,
@@ -30,6 +32,7 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.datatypes.tuning_job_stat
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.models.model_skipopt import (
     SkipOptimizationPredicate,
+    AlwaysSkipPredicate,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.models.model_transformer import (
     TransformerOutputModelFactory,
@@ -39,12 +42,26 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.models.model_transformer 
 from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.base_classes import (
     AcquisitionClassAndArgs,
     LocalOptimizer,
+    ScoringFunction,
+    SurrogateOutputModel,
+    unwrap_acquisition_class_and_kwargs,
+    CandidateGenerator,
+)
+from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algorithm import (
+    BayesianOptimizationAlgorithm,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algorithm_components import (
     NoOptimization,
+    RandomStatefulCandidateGenerator,
+    RandomFromSetCandidateGenerator,
+    DuplicateDetectorIdentical,
 )
-from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.common import (
+from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algorithm_components import (
+    IndependentThompsonSampling,
+)
+from syne_tune.optimizer.schedulers.searchers.utils.exclusion_list import (
     ExclusionList,
+    ExclusionListFromState,
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.defaults import (
     DEFAULT_NUM_INITIAL_CANDIDATES,
@@ -62,7 +79,6 @@ from syne_tune.optimizer.schedulers.searchers.utils.common import (
     ConfigurationFilter,
     Configuration,
 )
-from syne_tune.optimizer.schedulers.utils.simple_profiler import SimpleProfiler
 
 logger = logging.getLogger(__name__)
 
@@ -159,7 +175,6 @@ class ModelBasedSearcher(StochasticSearcher):
             skip_optimization=skip_optimization,
             state_converter=state_converter,
         )
-        self.set_profiler(model_factory_main.profiler)
         self._cost_attr = cost_attr
         self._resource_attr = resource_attr
         self._filter_observed_data = filter_observed_data
@@ -293,7 +308,7 @@ class ModelBasedSearcher(StochasticSearcher):
         def skip_all(config: Configuration) -> bool:
             return False
 
-        return ExclusionList(
+        return ExclusionListFromState(
             self.state_transformer.state,
             filter_observed_data=skip_all
             if skip_observed
@@ -337,8 +352,6 @@ class ModelBasedSearcher(StochasticSearcher):
         else:
             pick_random = True  # Initial configs count as random here
         if pick_random and config is None:
-            if self.do_profile:
-                self.profiler.start("random")
             for _ in range(GET_CONFIG_RANDOM_RETRIES):
                 _config = self._random_searcher.get_config()
                 if _config is None:
@@ -348,8 +361,6 @@ class ModelBasedSearcher(StochasticSearcher):
                 if not exclusion_candidates.contains(_config):
                     config = _config
                     break
-            if self.do_profile:
-                self.profiler.stop("random")
             # ``_random_searcher`` modified ``restrict_configurations``
             if self._restrict_configurations is not None:
                 self._restrict_configurations = (
@@ -365,18 +376,6 @@ class ModelBasedSearcher(StochasticSearcher):
         """
         start_time = time.time()
         state = self.state_transformer.state
-        if self.do_profile:
-            # Start new profiler block
-            skip_optimization = self.state_transformer.skip_optimization
-            if isinstance(skip_optimization, dict):
-                skip_optimization = skip_optimization[INTERNAL_METRIC_NAME]
-            meta = {
-                "fit_hyperparams": not skip_optimization(state),
-                "num_observed": state.num_observed_cases(),
-                "num_pending": len(state.pending_evaluations),
-            }
-            self.profiler.begin_block(meta)
-            self.profiler.start("all")
         # Initial configs come from ``points_to_evaluate`` or are drawn at random
         # We use ``exclusion_candidates`` even if ``allow_duplicates == True``, in order
         # to count how many unique configs have been suggested
@@ -416,9 +415,6 @@ class ModelBasedSearcher(StochasticSearcher):
             if cs_size is not None:
                 msg += f" Configuration space has size {cs_size}."
             logger.warning(msg)
-        if self.do_profile:
-            self.profiler.stop("all")
-            self.profiler.clear()
         self.cumulative_get_config_time += time.time() - start_time
 
         return config
@@ -461,10 +457,6 @@ class ModelBasedSearcher(StochasticSearcher):
         # ``random_state`` with this searcher here
         self._random_searcher = None
 
-    def set_profiler(self, profiler: Optional[SimpleProfiler]):
-        self.profiler = profiler
-        self.do_profile = profiler is not None
-
     @property
     def debug_log(self):
         return self._debug_log
@@ -487,3 +479,331 @@ class ModelBasedSearcher(StochasticSearcher):
                 restrict_configurations=self._restrict_configurations,
             )
             self._random_searcher.set_random_state(self.random_state)
+
+
+def create_initial_candidates_scorer(
+    initial_scoring: str,
+    model: SurrogateOutputModel,
+    acquisition_class: AcquisitionClassAndArgs,
+    random_state: np.random.RandomState,
+    active_metric: str = INTERNAL_METRIC_NAME,
+) -> ScoringFunction:
+    if initial_scoring == "thompson_indep":
+        if isinstance(model, dict):
+            assert active_metric in model
+            model = model[active_metric]
+        return IndependentThompsonSampling(model, random_state=random_state)
+    else:
+        acquisition_class, acquisition_kwargs = unwrap_acquisition_class_and_kwargs(
+            acquisition_class
+        )
+        return acquisition_class(
+            model, active_metric=active_metric, **acquisition_kwargs
+        )
+
+
+class BayesianOptimizationSearcher(ModelBasedSearcher):
+    """Common Code for searchers using Bayesian optimization
+
+    We implement Bayesian optimization, based on a model factory which
+    parameterizes the state transformer. This implementation works with
+    any type of surrogate model and acquisition function, which are
+    compatible with each other.
+
+    The following happens in :meth:`get_config`:
+
+    * For the first ``num_init_random`` calls, a config is drawn at random
+      (after ``points_to_evaluate``, which are included in the ``num_init_random``
+      initial ones). Afterwards, Bayesian optimization is used, unless there
+      are no finished evaluations yet (a surrogate model cannot be used with no
+      data at all)
+    * For BO, model hyperparameter are refit first. This step can be skipped
+      (see ``opt_skip_*`` parameters).
+    * Next, the BO decision is made based on
+      :class:`~syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algorithm.BayesianOptimizationAlgorithm`.
+      This involves sampling `num_init_candidates`` configs are sampled at
+      random, ranking them with a scoring function (``initial_scoring``), and
+      finally runing local optimization starting from the top scoring config.
+    """
+
+    def configure_scheduler(self, scheduler):
+        from syne_tune.optimizer.schedulers.scheduler_searcher import (
+            TrialSchedulerWithSearcher,
+        )
+
+        assert isinstance(
+            scheduler, TrialSchedulerWithSearcher
+        ), "This searcher requires TrialSchedulerWithSearcher scheduler"
+        super().configure_scheduler(scheduler)
+        # Allow model factory to depend on ``scheduler`` as well
+        model_factory = self.state_transformer.model_factory
+        if isinstance(model_factory, dict):
+            model_factories = list(model_factory.values())
+        else:
+            model_factories = [model_factory]
+        for model_factory in model_factories:
+            model_factory.configure_scheduler(scheduler)
+
+    def register_pending(
+        self, trial_id: str, config: Optional[Dict[str, Any]] = None, milestone=None
+    ):
+        """
+        Registers trial as pending. This means the corresponding evaluation
+        task is running. Once it finishes, update is called for this trial.
+        """
+        state = self.state_transformer.state
+        if not state.is_pending(trial_id):
+            assert not state.is_labeled(trial_id), (
+                f"Trial trial_id = {trial_id} is already labeled, so cannot "
+                "be pending"
+            )
+            self.state_transformer.append_trial(trial_id, config=config)
+
+    def _fix_resource_attribute(self, **kwargs):
+        pass
+
+    def _postprocess_config(self, config: Dict[str, Any]) -> Dict[str, Any]:
+        return config
+
+    def _create_random_generator(self) -> CandidateGenerator:
+        if self._restrict_configurations is None:
+            return RandomStatefulCandidateGenerator(
+                hp_ranges=self._hp_ranges_for_prediction(),
+                random_state=self.random_state,
+            )
+        else:
+            hp_ranges = self._hp_ranges_for_prediction()
+            if hp_ranges.is_attribute_fixed():
+                ext_config = {hp_ranges.name_last_pos: hp_ranges.value_for_last_pos}
+            else:
+                ext_config = None
+            return RandomFromSetCandidateGenerator(
+                base_set=self._restrict_configurations,
+                random_state=self.random_state,
+                ext_config=ext_config,
+            )
+
+    def _update_restrict_configurations(
+        self,
+        new_configs: List[Dict[str, Any]],
+        random_generator: RandomFromSetCandidateGenerator,
+    ):
+        # ``random_generator`` maintains all positions returned during the
+        # search. This is used to restrict the search of ``new_configs``
+        new_ms = set(
+            self.hp_ranges.config_to_match_string(config) for config in new_configs
+        )
+        num_new = len(new_configs)
+        assert num_new >= 1
+        remove_pos = []
+        for pos in random_generator.pos_returned:
+            config_ms = self.hp_ranges.config_to_match_string(
+                self._restrict_configurations[pos]
+            )
+            if config_ms in new_ms:
+                remove_pos.append(pos)
+                if len(remove_pos) == num_new:
+                    break
+        if len(remove_pos) == 1:
+            self._restrict_configurations.pop(remove_pos[0])
+        else:
+            remove_pos = set(remove_pos)
+            self._restrict_configurations = [
+                config
+                for pos, config in enumerate(self._restrict_configurations)
+                if pos not in remove_pos
+            ]
+
+    def _get_config_modelbased(
+        self, exclusion_candidates, **kwargs
+    ) -> Optional[Configuration]:
+        # Obtain current :class:`SurrogateModel` from state transformer. Based on
+        # this, the BO algorithm components can be constructed
+        # Note: Asking for the model triggers the posterior computation
+        model = self.state_transformer.model()
+        # Select and fix target resource attribute (relevant in subclasses)
+        self._fix_resource_attribute(**kwargs)
+        # Create BO algorithm
+        random_generator = self._create_random_generator()
+        initial_candidates_scorer = create_initial_candidates_scorer(
+            initial_scoring=self.initial_scoring,
+            model=model,
+            acquisition_class=self.acquisition_class,
+            random_state=self.random_state,
+        )
+        local_optimizer = self.local_minimizer_class(
+            hp_ranges=self._hp_ranges_for_prediction(),
+            model=model,
+            acquisition_class=self.acquisition_class,
+            active_metric=INTERNAL_METRIC_NAME,
+        )
+        bo_algorithm = BayesianOptimizationAlgorithm(
+            initial_candidates_generator=random_generator,
+            initial_candidates_scorer=initial_candidates_scorer,
+            num_initial_candidates=self.num_initial_candidates,
+            local_optimizer=local_optimizer,
+            pending_candidate_state_transformer=None,
+            exclusion_candidates=exclusion_candidates,
+            num_requested_candidates=1,
+            greedy_batch_selection=False,
+            duplicate_detector=DuplicateDetectorIdentical(),
+            sample_unique_candidates=False,
+            debug_log=self.debug_log,
+        )
+        # Next candidate decision
+        _config = bo_algorithm.next_candidates()
+        if len(_config) > 0:
+            config = self._postprocess_config(_config[0])
+            if self._restrict_configurations is not None:
+                # Remove ``config`` from ``_restrict_configurations``
+                self._update_restrict_configurations([config], random_generator)
+        else:
+            config = None
+        return config
+
+    def get_batch_configs(
+        self,
+        batch_size: int,
+        num_init_candidates_for_batch: Optional[int] = None,
+        **kwargs,
+    ) -> List[Configuration]:
+        """
+        Asks for a batch of ``batch_size`` configurations to be suggested. This
+        is roughly equivalent to calling ``get_config`` ``batch_size`` times,
+        marking the suggested configs as pending in the state (but the state
+        is not modified here). This means the batch is chosen sequentially,
+        at about the cost of calling ``get_config`` ``batch_size`` times.
+
+        If ``num_init_candidates_for_batch`` is given, it is used instead
+        of ``num_init_candidates`` for the selection of all but the first
+        config in the batch. In order to speed up batch selection, choose
+        ``num_init_candidates_for_batch`` smaller than
+        ``num_init_candidates``.
+
+        If less than ``batch_size`` configs are returned, the search space
+        has been exhausted.
+
+        Note: Batch selection does not support ``debug_log`` right now: make sure
+        to switch this off when creating scheduler and searcher.
+        """
+        assert round(batch_size) == batch_size and batch_size >= 1
+        configs = []
+        if batch_size == 1:
+            config = self.get_config(**kwargs)
+            if config is not None:
+                configs.append(config)
+        else:
+            # :class:`DebugLogWriter` does not support batch selection right now,
+            # must be switched off
+            assert self.debug_log is None, (
+                "``get_batch_configs`` does not support debug_log right now. "
+                + "Please set ``debug_log=False`` in search_options argument "
+                + "of scheduler, or create your searcher with ``debug_log=False``"
+            )
+            exclusion_candidates = self._get_exclusion_candidates(
+                skip_observed=self._allow_duplicates
+            )
+            pick_random = True
+            while pick_random and len(configs) < batch_size:
+                config, pick_random = self._get_config_not_modelbased(
+                    exclusion_candidates
+                )
+                if pick_random:
+                    if config is not None:
+                        configs.append(config)
+                        # Even if ``allow_duplicates == True``, we don't want to have
+                        # duplicates in the same batch
+                        exclusion_candidates.add(config)
+                    else:
+                        break  # Space exhausted
+            if not pick_random:
+                # Model-based decision for remaining ones
+                num_requested_candidates = batch_size - len(configs)
+                model = self.state_transformer.model()
+                # Select and fix target resource attribute (relevant in subclasses)
+                self._fix_resource_attribute(**kwargs)
+                # Create BO algorithm
+                random_generator = self._create_random_generator()
+                initial_candidates_scorer = create_initial_candidates_scorer(
+                    initial_scoring=self.initial_scoring,
+                    model=model,
+                    acquisition_class=self.acquisition_class,
+                    random_state=self.random_state,
+                )
+                local_optimizer = self.local_minimizer_class(
+                    hp_ranges=self._hp_ranges_for_prediction(),
+                    model=model,
+                    acquisition_class=self.acquisition_class,
+                    active_metric=INTERNAL_METRIC_NAME,
+                )
+                pending_candidate_state_transformer = None
+                if num_requested_candidates > 1:
+                    # Internally, if num_requested_candidates > 1, the candidates are
+                    # selected greedily. This needs model updates after each greedy
+                    # selection, because of one more pending evaluation.
+                    # We need a copy of the state here, since
+                    # ``pending_candidate_state_transformer`` modifies the state (it
+                    # appends pending trials)
+                    temporary_state = copy.deepcopy(self.state_transformer.state)
+                    pending_candidate_state_transformer = ModelStateTransformer(
+                        model_factory=self.state_transformer.model_factory,
+                        init_state=temporary_state,
+                        skip_optimization=AlwaysSkipPredicate(),
+                    )
+                bo_algorithm = BayesianOptimizationAlgorithm(
+                    initial_candidates_generator=random_generator,
+                    initial_candidates_scorer=initial_candidates_scorer,
+                    num_initial_candidates=self.num_initial_candidates,
+                    num_initial_candidates_for_batch=num_init_candidates_for_batch,
+                    local_optimizer=local_optimizer,
+                    pending_candidate_state_transformer=pending_candidate_state_transformer,
+                    exclusion_candidates=exclusion_candidates,
+                    num_requested_candidates=num_requested_candidates,
+                    greedy_batch_selection=True,
+                    duplicate_detector=DuplicateDetectorIdentical(),
+                    sample_unique_candidates=False,
+                    debug_log=self.debug_log,
+                )
+                # Next candidate decision
+                _configs = [
+                    self._postprocess_config(config)
+                    for config in bo_algorithm.next_candidates()
+                ]
+                configs.extend(_configs)
+                if self._restrict_configurations is not None:
+                    self._update_restrict_configurations(_configs, random_generator)
+        return configs
+
+    def evaluation_failed(self, trial_id: str):
+        # Remove pending evaluation
+        self.state_transformer.drop_pending_evaluation(trial_id)
+        # Mark config as failed (which means it will be blacklisted in
+        # future get_config calls)
+        self.state_transformer.mark_trial_failed(trial_id)
+
+    def _new_searcher_kwargs_for_clone(self) -> Dict[str, Any]:
+        """
+        Helper method for ``clone_from_state``. Args need to be extended
+        by ``model_factory``, ``init_state``, ``skip_optimization``, and others
+        args becoming relevant in subclasses only.
+
+        :return: kwargs for creating new searcher object in ``clone_from_state``
+        """
+        return dict(
+            config_space=self.config_space,
+            metric=self._metric,
+            clone_from_state=True,
+            hp_ranges=self.hp_ranges,
+            acquisition_class=self.acquisition_class,
+            map_reward=self.map_reward,
+            local_minimizer_class=self.local_minimizer_class,
+            num_initial_candidates=self.num_initial_candidates,
+            num_initial_random_choices=self.num_initial_random_choices,
+            initial_scoring=self.initial_scoring,
+            skip_local_optimization=self.skip_local_optimization,
+            cost_attr=self._cost_attr,
+            resource_attr=self._resource_attr,
+            filter_observed_data=self._filter_observed_data,
+            allow_duplicates=self._allow_duplicates,
+            restrict_configurations=self._restrict_configurations,
+        )
