+++ tst/schedulers/bayesopt/test_bo_algorithm_functions.py
@@ -16,8 +16,6 @@ from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algo
 )
 from syne_tune.optimizer.schedulers.searchers.bayesopt.tuning_algorithms.bo_algorithm_components import (
     NoOptimization,
-)
-from syne_tune.optimizer.schedulers.searchers.bayesopt.utils.duplicate_detector import (
     DuplicateDetectorIdentical,
 )
 from syne_tune.optimizer.schedulers.searchers.utils.hp_ranges_factory import (
